Add error field to Exercise model
If exercise is not found from Plussa anymore it is removed from
database or, if exercise is marked for grading, the error field is set
to True. This is shown to teacher in the exercise list view with a
badge and error message. Teacher is asked either to remove the
exercise or keep the exercise and ignore the warning.